requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
background==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.9
django-celery-results==2.5.0
django-compat==1.0.15
django-constance==2.9.1
django-crontab==0.7.1
django-picklefield==3.1
django4-background-tasks==1.2.7
idna==3.4
kombu==5.2.4
mutagen==1.46.0
paho-mqtt==1.6.1
prompt-toolkit==3.0.38
pygame==2.4.0
python-vlc==3.0.18121
pytz==2023.3
redis==4.5.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
wget==3.2
